Size is not grandeur, and territory does not make a nation

+++ 0x00-hello_world/6-size.c
@@ -11,11 +11,11 @@ int main(void)
 	long long int a;
 	float f;
 
-printf("Size of a char: %lu byte(s)\n", sizeof(x));
-printf("Size of an int: %lu byte(s)\n", sizeof(y));
-printf("Size of a long int: %lu bytes(s)\n", sizeof(z));
-printf("Size of a long long int: %lu bytes(s)\n", sizeof(a));
-printf("Size of a float: %lu bytes(s)\n", sizeof(f));
+printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(x));
+printf("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(y));
+printf("Size of a long int: %lu bytes(s)\n", (unsigned long)sizeof(z));
+printf("Size of a long long int: %lu bytes(s)\n", (unsigned long)sizeof(a));
+printf("Size of a float: %lu bytes(s)\n", (unsigned long)sizeof(f));
 return (0);
 }
